requirements.txt
keras==2.3.1
cv2==3.4.2
numpy==1.16.6
matplotlib==2.2.5
python==2.7.16
tkinter==8.5
Pillow==6.2.2

